All test passed. Need to test different platforms and review docs/code a bit, but mostly done
