+++ pytorch_widedeep/models/__init__.py
@@ -1,7 +1,6 @@
 from pytorch_widedeep.models.text import (
     HFModel,
     BasicRNN,
-    Transformer,
     AttentiveRNN,
     StackedAttentiveRNN,
 )
