+++ pytorch_widedeep/models/rec/__init__.py
@@ -2,3 +2,4 @@ from pytorch_widedeep.models.rec.din import DeepInterestNetwork
 from pytorch_widedeep.models.rec.deepfm import DeepFactorizationMachine
 from pytorch_widedeep.models.rec.deepffm import DeepFieldAwareFactorizationMachine
 from pytorch_widedeep.models.rec.xdeepfm import ExtremeDeepFactorizationMachine
+from pytorch_widedeep.models.rec.basic_transformer import Transformer
