+++ pytorch_widedeep/models/rec/deepffm.py
@@ -15,6 +15,11 @@ class DeepFieldAwareFactorizationMachine(BaseTabularModelWithAttention):
     systems. Adaptation of the paper Field-aware Factorization Machines in a
     Real-world Online Advertising System, Juan et al. 2017.
 
+    This class implements only the 'Deep' component of the model described in
+    the paper. The linear component is not implemented 'internally' and, if
+    one wants to include it, it can be easily added using the 'Wide'/linear
+    component in this library. See the examples in the examples folder.
+
     Note that in this case, only categorical features are accepted. This is
     because the embeddings of each feature will be learned using all other
     features. Therefore these embeddings have to be all of the same nature.
