+++ pytorch_widedeep/models/text/__init__.py
@@ -3,5 +3,4 @@ from pytorch_widedeep.models.text.rnns import (
     AttentiveRNN,
     StackedAttentiveRNN,
 )
-from pytorch_widedeep.models.text.miscellaneous import Transformer
 from pytorch_widedeep.models.text.huggingface_transformers import HFModel
