--- pytorch_widedeep/models/text/miscellaneous/__init__.py
@@ -1,1 +1,0 @@
-from pytorch_widedeep.models.text.miscellaneous.basic_transformer import Transformer
